Create diff tooling
Keeping track of patch files is difficult and time consuming.
Rather than check in only a cumulative patch for all files in
a project repository it would be better to check in the changed
files themselves.  The tooling can then clone the git project
and generate a patch file for each changed file.

The mkdiffs.py tool takes in a yaml configuration file that
specifies the projects to create patches for, the location
of the changes, and a location for the patch files.  The
tool can be run without any parameters to clone each project and
generate all the patch files, or with -s to skip the cloning,
assuming that step has been done previously.

Once the patch files have been generated, the developer can
pick the changed ones and copy them to the appropriate
directory.  This step is not automatic since each patch
file contains timestamps and would be shown as always changed.

Change-Id: Id23d8cb67e6ab009825f5728395b5ec85cfa6840