Power KVM repo build changes for swift
This change modifies the Power KVM builds to allow the repository
server to be either a Power Compute host or a Power Swift Host.

Power KVM repo packages to build for the ppc64le architecture is also
now variable and dependent on whether a Power Compute host or Power
Swift hosts are in the inventory.

Change-Id: I322f64657923c94da4a00d4149abadea2df45f16