Add playbook structure
This change set adds a playbook structure which
is currently intended to contain playbook(s) and roles
for pre-OSA configuration of OSA input files as well as
configuration of nodes such as disk prep for Swift.

The playbook is using host groups defined by the node
grouping in the inventory.yml file and will also use
the reference-architecture property in the inventory.yml
in conditionals.

Change-Id: Ia407cc922a34c40c6d6f7afecf8999d9ae04406c