Fix Swift ring division
This change set fixes a bug in the conversion of the
ring-disks properties in the inventory to the proper
groups in OpenStack Ansible user config.

After this change the generate_user_config code will
now allow account and container disks to be on separate
groups in the OSA inventory.  It will also allow all rings
to use the same set of disks.

Change-Id: Ia66c6492a2b948d210e6a82bca2e619812261e7c