Configure HA proxy for Swift standalone
This adds support for configuring the HA proxy service differently
when Swift is deployed without the private compute cloud.

The list of services that HA proxy is configured for is different
when the private compute cloud is not deployed.

OpenStack Ansible does not have an easily pluggable HA proxy
configuration which was noted on [1].  This is what causes
us to have two full HA proxy service config yamls.

The OpenStack Ansible haproxy-install playbook is conditionally
re-running the haproxy_server role to add the Keystone internal
service depending on an equality check on the internal and external
floating IPs.

Since the HA proxy configuration is being specified as an override
variable it takes precedence over the variables they are using on
the re-run.  This leads us to specify two separate files and do
the a similar check on the IP addresses.

[1] https://review.openstack.org/#/c/211209/

Change-Id: Ieeec99956a9124dd0c3c7cb99548f3af7cff681c