Patch ansible 1.9.4 to fix ansible_processor_vcpus.
This patch improves the way fact checking in ansible
parses /proc/cpuinfo.  The default value (1) will always
be used for both the cpu and virtual cpu count if this
patch is not applied.

This is backporting fixes from ansible v2.0 (already
upstream in the community) to the version of ansible we
currently use (v1.9.4).

This certainly impacts multiple openstack services.
One example is Swift.  This patch is needed
so that worker threads of the Swift processes can be
set correctly according to the number of CPUs the node has.

Change-Id: Ibcd743dc0beca31ca6b6d75a5e8665f8a144fa18