Replicate /etc/openstack_deploy on each controller node
This is done in a post deployment playbook.  This data provides topology
and secrets which will be needed for update and add a node.  OSA would
have to be installed on these other nodes.  The user would drive the
installation of OSA and new logic would have to be developed to consume
this pre-existing OSA meta data.

OSA related playbooks are placed in the osa directory.  Stuff in the root
directory provides orchestration and is common to all ulysses projects.

Change-Id: Ic3ea0ecf52c6695e0013025abfbeaa12462fde90