Replicate /etc/openstack_deploy on each controller node
This is done in a post deployment playbook.  This data provides topology
and secrets which will be needed for update and add a node.  OSA would
have to be installed on these other nodes.  The user would drive the
installation of OSA and new logic would have to be developed to consume
this pre-existing OSA meta data.

Change-Id: I4eed67092fcc454cc6020df5a3fd3b7d131b9add