Add arch test to os-setup-swift playbook
The os-setup-swift playbook is used to modify Swift configuration
or update the Swift code level without regenerating and syncing
the Swift keys or rings.

This playbook needs the same fix as the install playbook to fetch
the venv tgz using the architecture in the URL.  Without this fix
the playbook will pull the x86 venv for ppc64le systems.

Change-Id: Idbf6cedc3ba46efda5b083fcf73b2bfe032f2de5