Fix AIO builds
Add genesis inventory file for AIO which defines a single node
configuration.  The projects os-services and ceph-services do not
use this new inventory file as the code recognizes that this is
an AIO config, however, the opsmgr playbooks do use the new
inventory file along with OSA inventory.  This change provides
a common runtime environment for POST-OSA playbooks.

Change-Id: I11dfb27e4a41bd956f9c2ca50cd06e20cb1dd297