Swift proxy performance fixes
This is a copy of [1] which is in stable/mitaka but not yet
included in the tag level we use.

Performance testing of an OSA deployed Swift cluster hit several issues.
Comparison of the proxy conf with the Swift Auth doc [2] found
two settings that affect performance that that were not set in the template.

include_service_catalog - This defaults to True which makes the proxy server
fetch the service catalog on every request in the authtoken middleware.
This hammers Keystone.

cache - When using memcache this should be set so the authtoken middleware
uses the correct cache and not go back to Keystone on every request.

[1] https://review.openstack.org/#/c/392112/
[2] http://docs.openstack.org/developer/swift/overview_auth.html

Change-Id: I774dac10a30eab29272da297a3a60fc4bec28c04
(cherry picked from commit b7908439d3e68a55c844927d65f2370e15d026e4)