Use role_path instead of role_dir
The role_path is an ansible builtin.  The role_dir
was defined in the swift_hosts role for Swift disk
prep.  The role_dir was behaving differently in
OSA 13.3.5 than it was in OSA 13.1.0, but since it
is not needed we will remove it.

Change-Id: I8228153757331163ef8c6cc473a6d8fdb62ea3a0
(cherry picked from commit 02f64fe938a4203f0139d00b5c8ce6d1bf361701)