Enable openstack_hosts role without network_hosts
The openstack_hosts role will no longer generate
a playbook error when the network_hosts group (used
for the neutron openstack service) is not defined.

The generate_user_config.py will only define the
network_hosts group when the reference architecture
contains neutron (e.g. private-compute-cloud).

The playbook error was not present in OSA 13.1.0
but is present in OSA 13.3.5.  The injecting drop
is listed below.

  [1] https://github.com/openstack/openstack-ansible-openstack_hosts/commit/eb1b3afc7414cd25effd29335be7d6cc73c2a0ba

Change-Id: Ie0aa3f2804c14a7e0809c623806b4acd1ee6efbc