Remove Swift standalone haproxy config
This removes the haproxy service configuration variables file
and related code.

The OpenStack-Ansible commit 92eb98e1d2f367609640f9207b30038565a22339
changed the service configuration and haproxy playbook so that we do
not need to carry two separate files for keystone based on the internal
and external IP addresses.

Additionally, the upstream code has also been updated so that it only
writes the haproxy service configuration files if the service is
actually enabled to be deployed in the OpenStack Ansible run. This
allows us to not worry about the haproxy configuration and trying to
manage it in non-compute, Swift only clouds.

Change-Id: I9286cf821d2f8283386b08d7821977a2db9181fb