Set GENESIS_DIR to a full path not a relative path
The previous path for GENESIS_DIR was inside
the os-services directory.  The new path
will be in /opt/cluster-genesis by default.
This can be overriden by exporting the GENESIS_DIR
environment variable prior to calling bootstrap-cluster.sh.

This allows the disk prep code for Openstack Swift
(located in os-services) and the disk prep code
for Openstack Ceph (located in ceph-services)
to use the same dynamic inventory which is based
off the genesis inventory.  The dynamic inventory
is generated by cluster-genesis code which reads the
genesis inventory (located in /var/oprc/inventory.yml).

Change-Id: Ib62dcb0792e1af94eef450dd8022d9c6609d3491