Fix patch file name issue in versionDiff.sh
Patch file names created were changed to have "-" instead of
underscores in versionDiff.sh. But the patch file names that
are being applied to os-services/changes/ directory still have
"-"s. Fixing this issue in this patch.

Also fixing the path issue to the patch files when applying patches.
Issue is that the path for the patch files is becoming relative path
because of "dirname $0". Fixed this issue also in this patch.

Change-Id: I7dfaf2334c39258c5a2b2d7661103f4a29c60e84