Fix Swift disk prep for metadata nodes
The Swift disk preparation tasks fail on Swift metadata nodes.
The tasks are trying to update the inventory file with disk
information for the account, container, and object disks.

Swift metadata nodes do not have object disks and so the
object_devices variable ends up being defined and set to the empty
set.

The fix is to change the when conditional from "is defined" to
check set length.

Change-Id: Ibc8c731774982178f1d9c963c157c7b00f188d08
Closes-Bug: #1652331
(cherry picked from commit 1da37172834bc3eaacd543747aa343fddecc6d4e)