Disable ipv6 when containers use overlayfs
In the existing code, reloading of sysctl.conf is done in neutron
containers to make sure that ipv6 is disabled in those containers
even after restart of containers. This patch fixes an issue in
placing reload-sysctl-conf.sh in the containers in the correct
path for the case of containers using overlayfs also.

Change-Id: Ibd69f00b6637e7ae810832791a676d4d08912462