Support converged nodes by defining roles
Support converged nodes by allowing defining of roles in
node templates. This is backwards compatible in the way such that
the existing node-templates with the names controllers, compute, etc
are handled similar to how they are dealt with till now. This patch
allows user to add new node-template with his/her own new name and
user can specify roles under the roles section.

Roles allowed for now in this patch are controller, compute,
ceph-monitor, and swift-proxy. Role support is not added for
swift-metadata and swift-object services.

Change-Id: Id35ebd3a5a41efd2817ceef548a2a56fbf275fa2