Allow applying clean patches when porting oss
Earlier versionDiff.sh was allowing to apply all patches or none
when porting from one version of OpenStack-Ansible to another. With
this code change, versionDiff.sh allows user to choose a new option
to apply only those patches which can be applied to
os-services/changes/ without resulting in errors.

Also added a new script cmpDiffs.sh that helps in
1. comparing the patches generated using mkdiffs.py and the patches
available in the directory os-services/osa/diffs/
2. deciding which patches are to be modified in os-services/osa/diffs
as part of porting os-services from one version of OpenStack-Ansible
to another version.
3. validating if the porting of os-services from one version of
OpenStack-Ansible to another is complete or not.
4. copying the patch files from os-services/.diffs/ to
os-services/osa/diffs/ avoiding manual copy of patch files.

Change-Id: I47b46cf375f4c306b2347b59aa6905a5df1e388c