First drop of dbimage-builder for DBaaS
The code installs and configures diskimage-builder, trove, and
tripleo-image-elements and applies patches in a VM previously
created by the user.

The ocata version of trove is used for diskimage-builder as it
provides better integration with xenial.  In fact, it proves support
for images built on both trusty and xenial and this capability is
exploited by dbimage-builder via an environment variable.

dbimage-builder places public ssh keys for the account running
the tool, for the root user on the controller, and optionally
for a user specified cloud user.  These public keys are copied
into the image's authorized_keys file.

A command file is generated in the dibvm allowing the toolkit
user to re-run the last command manually for debug purposes.
A command log file is also generated.

Upon successful completion, the generated image and the log
file are copied back to toolkit.  The dibvm generates a qcow2
file as it is about 1/4 of the size of a raw image. This minimizes
transfer times. The qcow2 file is converted into raw format,
updated to the glance, and a datastore is created.

The tool may be run from os-services on the control node or from
another node that has access to the control node and the dibvm.
This is accomplished by specifying the IP address or hostname of
the controller via the environment variable DBIMAGE_CONTROLLER_IP.
The ip address of the dibvm is a command argument as it always
needs to be specified.

Change-Id: I68ae789671032e53111831e5584f834a3b20e57b
(cherry picked from commit 03d90bb1cd9141e16896e27b72c453ab3f668148)