Add support for configuration validation_rules
This is one of the places that needs to be updated when a config
parameter is deprecated.  Added debug statements to postgresql
config code to determine how config data is being overlaid.

The first part of the validation rules code that identifies whether
there is validation rules for the database needs to be run locally,
not on the controller, so code had to be shifted from the controller
playbook to the deployer playbook.  Similar change for the
validate-args-table playbook.  It has no dependency anymore on the
dibvm, so it is moved to the deployer as well.

Change-Id: Iccd3384d3e4a492555894b73ee31cc3c5ae1c148