Add var for environment used only inside playbooks
When global_environment_variables is set in user_variables.yml,
this installs environment settings in /etc/environment on all
hosts and containers. These remain in place after deployment is
complete.

This patch adds a similar variable deployment_environment_variables
that defines environment strings applied only while the playbooks
are running. They leave nothing behind on the hosts or containers.
This may be used, for example, for proxy settings required only
during deployment.

This patch is a backport of OpenStack-Ansible ocata patch
https://review.openstack.org/#/c/467659/1 at
OpenStack-Ansible newton level in os-services.

Change-Id: Ifca99f645445674077058c91eccd7c4f995a2b0e