Add environment settings to galera_client
The galera_client role is invoked by other openstack roles by
using a dependency in the meta section of the roles.  It doesn't
seem to pickup the environment setting of the invoking role
nor does it seem to pickup environment settings placed in the
meta yaml file.  For that reason the environment values are
set on each task needing the proxy settings.

Change-Id: I3bad4e5e3503472ad1cc0a259baae889681cfc38
(cherry picked from commit 2104c7f476b6cddcfbd6126aadc390062676ffc3)