Allow user to specify deployment env variables
Allow user to specify deployment environment variables in
inventory file as a dictionary variable 'deployment-environment'.

This patch generates the user variable
'deployment_environment_variables' which is a dictionary of
environment variables that will be used during deployment. New
user variable file user_var_deploy_env.yml is generated for this.

This patch also makes sure that the required ip addresses are there
in the environment variable 'no_proxy' if the env variable
'http_proxy' or 'https_proxy' is defined.

Change-Id: I4cd6ed16a7a960730c7d8bd98b48eed0173f9bc8