Provide support for mongodb 3.4 community and enterprise editions
Community edition has to be compiled and it takes ~3 hours to build.

Command respectively are:

./scripts/dbimage-make.sh -i 9.3.89.213 -d mongodb -c -k luke
./scripts/dbimage-make.sh -i 9.3.89.213 -d mongodb -e -k luke

Note you only end up with one datastore definition as the second
invocation overwrites the first.  One has to manually run trove-manage
commands to recreate the datastore for the community edition -- the
1st image produced above.

The code drop fixes a trove problem in the management of users
for mongodb.  Like other databases, users should be expressed as a
single unit, however, that is not the case with Trove mongo support.
It requires that the user be specified relative to a database.  This
results in a single user being listed multiple times if he is
associated with multiple databases.  This does not scale well and is
confusing as each database.user entity can be updated independently
to change the password.  There is only one user and one password.

This commit maintains compatibility with the database.user format
but allows users to be specified independently as long as an
initial database is provided.  The database is created separately,
but importantly the Trove UI now only shows the user once.

There is a Trove restriction in the management of databases.  It is
not possible to create an empty database in Trove.  Mongodb garbage
collects empty databases, so the Trove create database command and
UI does not work properly for mongo as the database does not persist.

To overcome this problem, the database must be created from the client.

1) Create the user in Trove.  For example, user=luke, initial db=spain
2) Create the database from a mongo client

mongo --host 9.3.80.128 --authenticationDatabase spain -u luke -p
- use spain
- db.createCollection("food")
- db.food.insert( { tapas: true } )
- exit

The database spain should be visible in Trove now.

Change-Id: I990b5ea380ee84ceccd64ecf371949e33a5c1b55