Limit datastore versions to valid images
The trove database can get in an inconsistent state if an image is
imported thru trove, but then deleted using the compute api. Trove
has no knowledge the image has been deleted. So in LaunchInstance
we now check that every datastore version has an image id that is
a valid image.

This adds one API call to the LaunchInstance workflow with no
noticeable performance difference.

Change-Id: Ic84abd8c34d28eabc602e8b231b5ba6c4546f433