Implement dbimage-upload.sh command
Command uploads a previously built image that was produced by
dbimage-make.sh.  This allows images to be reusable which saves
a lot of installation time and ensures that the same bits are
re-applied.  The image is mounted, updated, uploaded, and then
a datastore is created from the uploaded image.

Change-Id: I529071c4b2dbb57b88b73f4019fa8930f148ddf8
(cherry picked from commit 734434495d7fcae3f107c6422a33ebc0ba89be2b)