Add support for flavors.  Three new commands: show, change, upload
Includes playbooks and helper python script for parsing predefined
and customized values.  Predefined contains default values for the
five supported databases.  This file is not modified.  A second
file called customized is used to hold user modified values.

Also moved all helper scripts to helper directory.  The scripts
directory now only holds commands that are to be invoked by the
user.  This is cleaner approach from an end user perspective.

Change-Id: I0802470d44d8fa521c830ed6a5acbef9f6f4a78b