Fix mysql backup
The version of percona-xtrabackup provided by Xenial does not
work with the version of mysql provided by Xenial.  The percona
community does not provide the requisite package for ppc64le,
so support is being added to compile it.

There is also a new sys database that is added with mysql 5.7.  We
add it to the list of databses to be ignored, so that the user
cannot delete it.

Change-Id: I3e43305ae646aa1fa68ce732c062c8998b23ae46
(cherry picked from commit eb03737a7bfde052168e82c9ed61e462b2da7dfb)