Isolate DIB VM since it can't be rebooted
This is a diskimage-builder bug.  Clarify deployment model in
README.  Mandate dedicated deployer and dibvm nodes.  The
deployer node carries images and log files which have value.
The DIB VM has no persistent state.  It simply provides a
build service to the end user.  It is useful to the development
team though in that it provides a development environment
for providing fixes and support for new databases.

Make the toolset so that it includes non-privileged commands
only with the caveat that promptless sudo access be provided.
To accomplish this, ansible had to be installed through pip
so that local ansible cache could be managed consistently.
This brought in a newer version of ansible, so the playbooks
had to be updated to eliminate all warning messages.  The
ansible version is pinned at 2.3.2.

Change-Id: I15f631619f6c621275164770e1d72c52e3564e55
(cherry picked from commit 91e936bf0f2ebdb1bb8ac8c2231e992e1a95d0d1)