Fix postgresql 9.6
There was a mismatch between ubuntu and postgresql community regarding the content
of package postgresql.  For the former it is postgresql 9.6 for the latter postgresql 10.0.
This caused two versions of the database to be installed.  The fix is to refer to
these packages by version.

Also a config file value was deprecated in postgresql 9.6.  wal_level=hot_standby
is replaced by wal_level=replica.

Change-Id: I00a521052d72356a71baf34e5f09a0cd868b1a94