Allow redis-cli to bind to redis 3.2 datastore
The default is secure mode which only allows local instances
to connect via 127.0.0.1.  This was introduced in redis 3.2.6.
Our redis implementation is 3.0.6 and 3.2.8, so the trove
guestagent has to be aware of these differences.  Unfortunately,
the trove guest agent is not aware of the datastore version,
so the extra parameters (--protected-mode no) are passed into
the redis-server daemon via the systemd service.  This is done
in the integration element for redis.  This is analogous to
bind 0.0.0.0.

This fix is implemented on top of master branch version of
the trove guest agent for redis which includes a fix allowing
the client to set the password and hide the config command.

Change-Id: I388b74f13337216b3eaccc57c00a4612e1dac407